Rename jupytercad-schema to schema
