+++ packages/base/src/panelview/objecttree.tsx
@@ -18,7 +18,7 @@ import {
 
 import visibilitySvg from '../../style/icon/visibility.svg';
 import visibilityOffSvg from '../../style/icon/visibilityOff.svg';
-import { IJCadModel, IJCadObject } from '@jupytercad/jupytercad-schema';
+import { IJCadModel, IJCadObject } from '@jupytercad/schema';
 import {
   IControlPanelModel,
   IDict,
