+++ packages/jupytercad-worker/package.json
@@ -34,7 +34,7 @@
     },
     "dependencies": {
         "@jupytercad/jupytercad-opencascade": "^0.3.3",
-        "@jupytercad/jupytercad-schema": "^0.3.3",
+        "@jupytercad/schema": "^0.3.3",
         "uuid": "^8.3.2"
     },
     "devDependencies": {
