+++ packages/jupytercad-worker/src/types.ts
@@ -14,7 +14,7 @@ import {
   ISketchObject,
   ISphere,
   ITorus
-} from '@jupytercad/jupytercad-schema';
+} from '@jupytercad/schema';
 
 export interface IDict<T = any> {
   [key: string]: T;
