+++ packages/schema/src/schema/box.json
