+++ packages/schema/src/schema/cone.json
