+++ packages/schema/src/schema/extrusion.json
