+++ packages/schema/src/schema/fuse.json
