+++ packages/schema/src/schema/geomCircle.json
