+++ packages/schema/src/schema/geomLineSegment.json
