+++ packages/schema/src/schema/intersection.json
