+++ packages/schema/src/schema/placement.json
