+++ packages/schema/src/schema/sketch.json
