+++ packages/schema/src/schema/sphere.json
