+++ python/jupytercad-lab/package.json
@@ -56,8 +56,8 @@
         "@jupyter/docprovider": "^1.0.0",
         "@jupyter/ydoc": "^0.3.4 || ^1.0.2",
         "@jupytercad/jupytercad-opencascade": "^0.3.3",
-        "@jupytercad/jupytercad-schema": "^0.3.3",
         "@jupytercad/jupytercad-worker": "^0.3.3",
+        "@jupytercad/schema": "^0.3.3",
         "@jupyterlab/application": "^4.0.0",
         "@jupyterlab/apputils": "^4.0.0",
         "@jupyterlab/coreutils": "^6.0.0",
